# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "gocommerce",
    "version": "0.1.0",
    "description": "A small, composable commerce engine for Go.\n\nMoney is always an integer count of the currency's minor unit plus its code — never a float, never a formatted string. The number of decimal places belongs to the currency (USD has 2, JPY has 0, KWD has 3) and formatting belongs to the client.\n\nEvery response is `{\"data\": ...}` or `{\"error\": {\"code\", \"message\"}}`. Every collection paginates.\n\nGuest checkout is permanent: a shopper needs a cart token and an email, never an account.",
    "license": {
      "name": "TBD"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This store"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/api/carts": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Open a cart",
        "description": "Returns the cart id, which is an unguessable token. Possessing it is the authorisation — there is no account, and there never has to be.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/CartResponse"
          }
        }
      }
    },
    "/api/carts/{cartId}": {
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Get a cart",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CartResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/carts/{cartId}/line-items": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Add a variant to the cart",
        "description": "Adding records the price now. Checkout is what makes a price authoritative, and it refuses rather than silently repricing.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "variant_id"
                ],
                "properties": {
                  "variant_id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CartResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/carts/{cartId}/line-items/{lineId}": {
      "patch": {
        "tags": [
          "cart"
        ],
        "summary": "Change a line's quantity",
        "description": "A quantity of zero removes the line, which is what a stepper stepping down to nothing means.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartID"
          },
          {
            "$ref": "#/components/parameters/LineID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "quantity"
                ],
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CartResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "cart"
        ],
        "summary": "Remove a line",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartID"
          },
          {
            "$ref": "#/components/parameters/LineID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CartResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/checkout/{code}": {
      "post": {
        "tags": [
          "checkout"
        ],
        "summary": "Check out with a payment method",
        "description": "Validates the cart, reserves inventory and creates the order in one transaction, then starts payment outside it — a gateway having a slow day must not become the store having a slow day.\n\nIf a line's price changed, its product went inactive, or stock ran out, the whole checkout is refused with 409 and per-line detail; the cart is refreshed to current prices so the shopper can confirm again.\n\nSend an `Idempotency-Key` header. The same key replays the original result instead of creating a second order; if the first attempt created the order but payment initiation failed, a retry resumes payment on that same order.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "description": "A payment method code, such as `cod`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Lang"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The order, and what the client must do to pay.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "order": {
                          "$ref": "#/components/schemas/Order"
                        },
                        "payment": {
                          "$ref": "#/components/schemas/PaymentIntent"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "The cart is no longer valid at these prices.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "example": {
                    "error": {
                      "code": "conflict",
                      "message": "the cart is no longer valid at these prices",
                      "details": [
                        {
                          "variant_id": 42,
                          "sku": "TEE-M-BLK",
                          "reason": "price_changed",
                          "current_price_minor": 2499
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/checkout/{code}/webhook": {
      "post": {
        "tags": [
          "checkout"
        ],
        "summary": "Gateway callback",
        "description": "A core-owned route that hands the raw request body to the payment module, so a gateway gets a documented callback URL without the module mounting anything itself. Signature verification is the module's job.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "*/*": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted by the provider."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/orders/{number}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Look up your own order",
        "description": "Guest checkout means there is no account to log into, so the order's access token — returned once at checkout — is the credential. A wrong token is reported as not found, so this cannot be used to discover which order numbers exist.",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OrderResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/admin/variants/{id}/inventory": {
      "post": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "Move stock",
        "security": [
          {
            "adminToken": []
          }
        ],
        "description": "Send `adjust` to move the count by a delta (receiving stock), or `set` to replace it (a stock take). Neither may leave on-hand below the quantity already reserved at that location for orders in flight.\n\n`location_id` says which shelf; omit it for the default one, which is what a single-location store always means.",
        "parameters": [
          {
            "$ref": "#/components/parameters/PathID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "adjust": {
                    "type": "integer",
                    "example": 25
                  },
                  "set": {
                    "type": "integer",
                    "example": 100
                  },
                  "location_id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Omit for the default location."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The variant.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Variant"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/admin/inventory/low-stock": {
      "get": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "Variants at or below a stock threshold",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 5,
              "minimum": 0
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/PageNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Variants running out.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Variant"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ListMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/orders/{id}": {
      "get": {
        "tags": [
          "admin:orders"
        ],
        "summary": "Get an order",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OrderResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "admin:orders"
        ],
        "summary": "Correct an order's contact details and payment record",
        "description": "An email with a typo in it means the customer never hears from the shop again; an address with the wrong house number means the parcel goes to the wrong door; a payment reference is what somebody reconciles against a bank statement. None of them is a state change, and each is a field a person typed once.\n\nCorrecting an address does not move a parcel that has already gone out.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OrderResponse"
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Either a static token from Config.AdminTokens — the right credential for a script, since it has no session and no expiry — or a superuser session token from /api/admin/auth-with-password, which is the right credential for a person. Both arrive the same way and satisfy the same middleware."
      }
    },
    "parameters": {
...
      "CartID": {
        "name": "cartId",
        "in": "path",
        "required": true,
        "description": "The cart's opaque token.",
        "schema": {
          "type": "string"
        }
      },
      "LineID": {
        "name": "lineId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "No such resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "CartResponse": {
        "description": "The cart.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}